restructuring; put many dto's into inner classes
