+++ src/main/java/de/sakpaas/backend/dto/StatusDto.java
@@ -4,8 +4,6 @@ import lombok.Data;
 
 @Data
 public class StatusDto {
-
     private boolean status;
     private String version;
-
 }
