+++ src/main/java/de/sakpaas/backend/service/LocationApiSearchDAS.java
@@ -1,7 +1,6 @@
 package de.sakpaas.backend.service;
 
-import de.sakpaas.backend.dto.osmresult.OMSResultLocationDto;
-import de.sakpaas.backend.dto.osmresult.OSMResultLocationListDto;
+import de.sakpaas.backend.dto.OSMResultLocationListDto;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestTemplate;
@@ -13,8 +12,8 @@ import static java.util.Collections.emptyList;
 @Component
 public class LocationApiSearchDAS {
 
-    public List<OMSResultLocationDto> getLocationByCoordinates(Double latitude, Double longitude,
-                                                               Double radius) {
+    public List<OSMResultLocationListDto.OMSResultLocationDto> getLocationByCoordinates(Double latitude, Double longitude,
+                                                                                        Double radius) {
 
         final String url = "https://overpass-api.de/api/interpreter?data=[out:json];" +
                 "node[shop=supermarket](around:" + radius.toString() + "," + latitude.toString() + "," + longitude.toString() + ");out;" +
@@ -38,7 +37,7 @@ public class LocationApiSearchDAS {
      * @param countryCode currently Dummy as lookup for area id is missing
      * @return list of supermarkets in Country
      */
-    public List<OMSResultLocationDto> getLocationsForCountry(String countryCode) {
+    public List<OSMResultLocationListDto.OMSResultLocationDto> getLocationsForCountry(String countryCode) {
         final String url = "https://overpass-api.de/api/interpreter?data=[out:json][timeout:2500];area(3600051477)->." +
                 "searchArea;(node[shop=supermarket](area.searchArea);way[shop=supermarket](area.searchArea);" +
                 "node[shop=chemist](area.searchArea);way[shop=chemist](area.searchArea););out center;";
@@ -52,7 +51,7 @@ public class LocationApiSearchDAS {
         return response.getBody().getElements();
     }
 
-    public OMSResultLocationDto getLocationById(Long id) {
+    public OSMResultLocationListDto.OMSResultLocationDto getLocationById(Long id) {
         final String url = "https://overpass-api.de/api/interpreter?data=[out:json];node(" + id + ");out;";
 
         RestTemplate restTemplate = new RestTemplate();
